Separate the represented combat responsibility from the Session root (#617)
Eleventh delivery of the refactor push. 66 methods move into 5 bounded
submodules under session/combat: state, melee, damage, vitals and death.
Every file is under the 1,000-line review budget.

session/mod.rs drops from 37,808 to 36,338 production lines. Across the eleven
deliveries the root falls from 75,841 to 36,338 - 52 per cent retired into
navigable owners.

The general combat-state module is named state.rs rather than combat.rs, so
the path does not read session::combat::combat, matching the choice made for
movement in #603.

Methods that mention damage or death but belong to spell, aura, creature,
gameobject, quest, pet, loot or item responsibilities already moved with those
families - creature kills in particular went with #599 - and are not touched
again.

Safety established by comparison: the impl surface is 966 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.
