+++ crates/wow-world/src/session/combat/mod.rs
@@ -1,0 +1,10 @@
+//! Represented combat responsibility, separated from the
+//! Session root under #617. Each submodule owns one complete operation
+//! group; the canonical owners keep authority over the state they touch.
+
+use super::*;
+mod damage;
+mod death;
+mod melee;
+mod state;
+mod vitals;
